+++ example_websocket_api_orders.py
@@ -36,8 +36,8 @@ import asyncio
 import logging
 import os
 
-api_key = ""
-api_secret = ""
+
+symbol = "BTCUSDT"
 
 
 async def binance_stream(ubwa):
@@ -45,26 +45,13 @@ async def binance_stream(ubwa):
         print(f"received data:\r\n{data}\r\n")
 
     api_stream = ubwa.create_stream(api=True, api_key=api_key, api_secret=api_secret,
-                                    stream_label="Bobs Websocket API",
                                     process_stream_data=handle_socket_message)
-    print(f"Start:")
-    ubwa.api.get_server_time(stream_id=api_stream)
-    ubwa.api.get_account_status(stream_id=api_stream)
-    orig_client_order_id = ubwa.api.create_order(stream_id=api_stream, price=1.0, order_type="LIMIT",
-                                                 quantity=15.0, side="SELL", symbol="BUSDUSDT")
+    ubwa.api.get_open_orders(stream_id=api_stream, symbol=symbol)
+#    ubwa.api.cancel_open_orders(stream_id=api_stream, symbol=symbol)
     ubwa.api.create_test_order(stream_id=api_stream, price=1.2, order_type="LIMIT",
-                               quantity=12.0, side="SELL", symbol="BUSDUSDT")
-    ubwa.api.ping(stream_id=api_stream)
-    ubwa.api.get_exchange_info(stream_id=api_stream, symbols=['BUSDUSDT'])
-    ubwa.api.get_order_book(stream_id=api_stream, symbol="BUSDUSDT", limit=2)
-    ubwa.api.cancel_order(stream_id=api_stream, symbol="BUSDUSDT", orig_client_order_id=orig_client_order_id)
-    ubwa.api.get_open_orders(stream_id=api_stream, symbol="BUSDUSDT")
-    ubwa.api.cancel_open_orders(stream_id=api_stream, symbol="BUSDUSDT")
-    ubwa.api.get_order(stream_id=api_stream, symbol="BUSDUSDT", orig_client_order_id=orig_client_order_id)
-
-    print(f"Finished! Waiting for responses:")
+                               quantity=12.0, side="SELL", symbol=symbol)
+    
     await asyncio.sleep(5)
-
     print(f"Stopping!")
     ubwa.stop_manager()